Extract TiFlash old start together
