+++ pkg/tidbver/tidbver.go
@@ -72,9 +72,9 @@ func TiFlashNotNeedSomeConfig(version string) bool {
 	return semver.Compare(version, "v5.4.0") >= 0 || strings.Contains(version, "nightly")
 }
 
-// TiFlashSupportRunWithoutConfig return true if the given version of TiFlash could be started without
-// a config file.
-func TiFlashSupportRunWithoutConfig(version string) bool {
+// TiFlashPlaygroundNewStartMode return true if the given version of TiFlash could be started
+// using the new implementation in TiUP playground
+func TiFlashPlaygroundNewStartMode(version string) bool {
 	return semver.Compare(version, "v7.1.0") >= 0 || strings.Contains(version, "nightly")
 }
